+++ GeneticBots/src/geneticbots/Target.java
@@ -4,7 +4,7 @@ import java.awt.Color;
 
 public class Target extends Actor
 {    
-    public Target(int x, int y)
+    public Target(double x, double y)
     {
         super(x,y,Color.RED,20);
     }
